// platform/docs-site/src/types/ast.ts
interface ParentNode extends Node {
  children: ASTNode[];
}
// ...
interface BaseFieldNode extends ParentNode {
  name: string;
  label?: string;
}

interface FieldNode extends BaseFieldNode {
  type: 'field';
}

interface FieldListNode extends BaseFieldNode {
  type: 'field_list';
  children: FieldNode[];
}